GLOB-20260 Remove winston dependency.
Why? Winston includes a number of dependencies for features we don't
use. This needlessly increases the surface area of projects that use it.

This PR is a wip to explore how we could remove the winston dependency.
It still needs more tests and more experimentation in an actual project.

I based my approach to moving winston off of how the very simple morgan
(https://github.com/expressjs/morgan) works. Basically morgan is just a
very lightweight wrapper on the writable stream that is passed to its
constructor. I'm trying the same thing for our logging with console and
loggly streams.